updated inventory

+++ src/api/bottler.py
@@ -24,8 +24,6 @@ def post_deliver_bottles(potions_delivered: list[PotionInventory], order_id: int
         num_potions = len(potions_delivered)
         connection.execute(sqlalchemy.text(f"UPDATE global_inventory SET num_green_potions = '{num_potions}'"))
         connection.execute(sqlalchemy.text("UPDATE global_inventory SET num_green_ml = '0'"))
-
-        
     
     print(f"potions delievered: {potions_delivered} order_id: {order_id}")
 
@@ -49,7 +47,7 @@ def get_bottle_plan():
             print(row)
             num_green_ml = row[0]
         print(num_green_ml)
-        num_potions = num_green_ml/100
+        num_potions = (int) (num_green_ml/100)
 
     return [
             {
